fix: refactor is in — statistics page is now a thin container, three chart components split out, useNowTick + useIsAdmin hooks replace the eslint-disable and the duplicated admin check. csv builder has proper jsdoc.
