+++ client/src/pages/ProfilePage.jsx
@@ -10,7 +10,8 @@ import {
 import {
   fetchManageableActivities,
 } from '../services/ManageActivitiesService.js'
-import { MANAGER_ROLES, ROLES } from '../constants/roles.js'
+import { MANAGER_ROLES } from '../constants/roles.js'
+import { useIsAdmin } from '../hooks/useIsAdmin.js'
 import Card from '../components/ui/Card.jsx'
 import Button from '../components/ui/Button.jsx'
 import Badge, {
@@ -36,9 +37,9 @@ export default function ProfilePage() {
 
   // Statistics dashboard is ADMIN-only — strictly narrower than
   // canManage because the backend endpoint is ADMIN-gated.
-  // Computed alongside canManage so both role checks live in
-  // the same spot and can't drift apart.
-  const isAdmin = user?.role === ROLES.ADMIN
+  // Sourced from the useIsAdmin hook so the role check stays
+  // consistent with any other page that asks the same question.
+  const isAdmin = useIsAdmin()
 
   // ── State ───────────────────────────────────────────────
